+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Fixed the performance of loading plans. ([#322](https://github.com/craftcms/commerce-stripe/issues/322))
+- Fixed the performance of loading plans in the control panel. ([#322](https://github.com/craftcms/commerce-stripe/issues/322))
+- Fixed a PHP error that could occur when handling a webhook request.
 
 ## 5.0.5 - 2025-03-02