+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fixed the performance of loading plans in the control panel. ([#322](https://github.com/craftcms/commerce-stripe/issues/322))
 - Fixed a PHP error that could occur when handling a webhook request.
+- Fixed a bug where choosing bank transfer as a payment method wouldn’t complete an order. ([#315](https://github.com/craftcms/commerce-stripe/issues/315))
+- Added `craft\commerce\stripe\SubscriptionGateway::handleCustomerCashBalanceTransaction()`.
+- Added `craft\commerce\stripe\SubscriptionGateway::transactionSupportsRefund()`.
+
 
 ## 5.0.5 - 2025-03-02