+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed a bug where checkout session completion would not mark transactions as successful from webhooks. ([#318](https://github.com/craftcms/commerce-stripe/issues/318))
 - Fixed a bug where choosing bank transfer as a payment method wouldn’t complete an order. ([#315](https://github.com/craftcms/commerce-stripe/issues/315))
+- Fixed a PHP error that could occur when handling a webhook request.
 - Added `craft\commerce\stripe\SubscriptionGateway::handleCustomerCashBalanceTransaction()`.
 - Added `craft\commerce\stripe\SubscriptionGateway::transactionSupportsRefund()`.